Set back the maximum capacity to uint32_t type as even on a 1 gb model in a 3-gram level we already reach some 65535 values for a word.

+++ inc/W2COrderedArrayTrie.hpp
@@ -133,7 +133,7 @@ namespace uva {
                  * @param cio the context index offset for computing the next contex index.
                  */
                 template<typename ARRAY_ELEM_TYPE>
-                class WordDataEntry : public ADynamicStackArray<ARRAY_ELEM_TYPE, uint16_t> {
+                class WordDataEntry : public ADynamicStackArray<ARRAY_ELEM_TYPE, uint32_t> {
                 public:
                     TShortId cio;
                 };
